+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Added *Copy Object Modifiers* node.
 - Added *Export Headers* option to the setup script.
 - Added *Indices* output to *Get Random List Elements* node.
+- Added *Overlay* mode to *Mix Falloff* node.
 
 ### Fixed